pyproject.toml
[tool.poetry]
name        = "poster"
version     = "0.1.0"
description = "A little app to create album's posters"
authors     = ["henrique <henrique@xonha.space>"]
license     = "MIT"
readme      = "README.md"
packages    = [{ include = "src" }]

[tool.poetry.scripts]
dev = "debug:main"
app = "run:main"

[tool.poetry.dependencies]
python        = "^3.11"
fastapi       = "^0.109.0"
uvicorn       = "^0.25.0"
jinja2        = "^3.1.3"
pytailwindcss = "^0.2.0"
python-dotenv = "^1.0.0"

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
